+++ src/hash/duncecap.c
@@ -142,7 +142,7 @@ duncecap_init_size(duncecap_t *self, char size)
     duncecap_store_t *store;
     uint64_t          len;
 
-    if (size > (sizeof(intptr_t) * 8)) {
+    if (((size_t)size) > (sizeof(intptr_t) * 8)) {
         abort();
     }
