+++ src/hash/hihat-a.c
@@ -94,7 +94,7 @@ hihat_a_init_size(hihat_t *self, char size)
     hihat_store_t *store;
     uint64_t       len;
 
-    if (size > (sizeof(intptr_t) * 8)) {
+    if (((size_t)size) > (sizeof(intptr_t) * 8)) {
 	abort();
     }
