+++ src/hash/tiara.c
@@ -90,7 +90,7 @@ tiara_init_size(tiara_t *self, char size)
     tiara_store_t *store;
     uint64_t       len;
 
-    if (size > (sizeof(intptr_t) * 8)) {
+    if (((size_t)size) > (sizeof(intptr_t) * 8)) {
 	abort();
     }
